# library-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the back-end API for an online library management system",
        "title": "Library Management API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/books": {
            "get": {
                "description": "获取图书列表，支持按关键词搜索 (Retrieve all books in the library, optionally filter by keyword)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理 (Book Management)"
                ],
                "summary": "获取所有图书 (Retrieve all books)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词 (Search keyword)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应 (Success Response)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Book"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误 (Server error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "将新图书添加到数据库中 (Add a new book with full information)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理 (Book Management)"
                ],
                "summary": "添加图书 (Add a new book to library)",
                "parameters": [
                    {
                        "description": "图书信息 (Book Information)",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "添加成功响应 (Success Response)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求数据错误 (Invalid request body)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "数据库错误 (Database error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "description": "根据图书 ID 获取详细信息 (Retrieve detailed information of a specific book by ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理 (Book Management)"
                ],
                "summary": "获取单本图书信息 (Retrieve book details by ID)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "图书 ID (Book ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应 (Success Response)",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    },
                    "400": {
                        "description": "无效 ID (Invalid book ID)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "图书未找到 (Book not found)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误 (Server error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "根据图书 ID 更新对应信息 (Update book details by ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理 (Book Management)"
                ],
                "summary": "更新图书 (Update book by ID)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "图书 ID (Book ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新后的图书信息 (Updated Book Information)",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功 (Success Response)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "无效 ID 或数据错误 (Invalid book ID or body)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "图书未找到 (Book not found)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "数据库错误 (Database error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "根据图书 ID 删除图书信息 (Delete a book record by its ID)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理 (Book Management)"
                ],
                "summary": "删除指定图书 (Delete a book by ID)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "图书 ID (Book ID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功 (Success Response)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "无效 ID (Invalid book ID)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "图书未找到 (Book not found)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误 (Server error)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string",
                    "example": "John Doe"
                },
                "available_copies": {
                    "description": "可借数量",
                    "type": "integer",
                    "example": 3
                },
                "genre": {
                    "description": "类型",
                    "type": "string",
                    "example": "Programming"
                },
                "id": {
                    "description": "图书 ID",
                    "type": "integer",
                    "example": 101
                },
                "isbn": {
                    "description": "ISBN",
                    "type": "string",
                    "example": "9781234567890"
                },
                "language": {
                    "description": "语言",
                    "type": "string",
                    "example": "English"
                },
                "shelf_number": {
                    "description": "书架号",
                    "type": "string",
                    "example": "A-12"
                },
                "title": {
                    "description": "标题",
                    "type": "string",
                    "example": "Golang Basics"
                }
            }
        },
...
    }
}